+++ src/componentStyles.scss
@@ -51,7 +51,6 @@
 @use 'components/FavoritingTileButton/favoritingTileButton';
 @use 'components/Toast/toast';
 @use 'components/ComposedModal/composedModal';
-@use 'components/BottomSheet/bottomSheet';
 
 // Patterns
 @use 'patterns/HeroBanner/heroBanner';
