--- src/components/BottomSheet/index.ts
@@ -1,1 +1,0 @@
-export { default as BottomSheet, type BottomSheetProps } from './BottomSheet';
