+++ src/components/Drawer/Drawer.tsx
@@ -1,10 +1,10 @@
+import * as Dialog from '@radix-ui/react-dialog';
+import classnames from 'classnames';
 import React from 'react';
 import { getCommonProps, noOp } from '../../utils';
-import classnames from 'classnames';
-import * as Dialog from '@radix-ui/react-dialog';
-import IconButton from '../IconButton/IconButton';
 import { ButtonVariants } from '../Button/types';
 import { Icon } from '../Icon';
+import IconButton from '../IconButton/IconButton';
 
 // You'll need to change the HTMLDivElement to match the top-level element of your component
 export interface DrawerProps extends React.HTMLAttributes<HTMLDivElement> {
@@ -20,6 +20,10 @@ export interface DrawerProps extends React.HTMLAttributes<HTMLDivElement> {
    * The content of the drawer
    */
   children?: React.ReactNode;
+  /**
+   * Drawer optional variant
+   */
+  variant?: 'bottomSheet' | 'default';
 }
 /**
  * ## Overview
@@ -27,9 +31,16 @@ export interface DrawerProps extends React.HTMLAttributes<HTMLDivElement> {
  * A component for displaying a drawer.
  *
  */
-const Drawer = ({ className, isOpen = false, onClose = noOp, children, ...props }: DrawerProps) => {
-  const { className: baseClassName, ...commonProps } = getCommonProps(props, 'Drawer');
-
+const Drawer = ({
+  className,
+  isOpen = false,
+  onClose = noOp,
+  children,
+  variant = 'default',
+  ...props
+}: DrawerProps) => {
+  const isBottomSheet = variant === 'bottomSheet';
+  const { className: baseClassName, ...commonProps } = getCommonProps(props, isBottomSheet ? 'BottomSheet' : 'Drawer');
   return (
     <Dialog.Root
       open={isOpen}
@@ -49,15 +60,17 @@ const Drawer = ({ className, isOpen = false, onClose = noOp, children, ...props 
           <Dialog.Title />
           <Dialog.Description />
           <Dialog.Close asChild>
-            <IconButton
-              onClick={onClose}
-              className={classnames(`${baseClassName}__close`)}
-              aria-label="Close"
-              data-testid="drawer-close"
-              variant={ButtonVariants.tertiary}
-            >
-              <Icon icon="CloseX" color="currentColor" />
-            </IconButton>
+            {!isBottomSheet && (
+              <IconButton
+                onClick={onClose}
+                className={classnames(`${baseClassName}__close`)}
+                aria-label="Close"
+                data-testid="drawer-close"
+                variant={ButtonVariants.tertiary}
+              >
+                <Icon icon="CloseX" color="currentColor" />
+              </IconButton>
+            )}
           </Dialog.Close>
           {children}
         </Dialog.Content>
