+++ src/patterns/InPlaceFilters/InPlaceFilter.stories.tsx
@@ -1,6 +1,7 @@
 import { Meta } from '@storybook/react';
 import { useState } from 'react';
-import InPlaceFilters, { FilterDropdownProps } from './InPlaceFilters';
+import InPlaceFilters from './InPlaceFilters';
+import { FilterDropdownProps } from './types';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
 const meta = {
